Switch from the bb5 tag to bb5_map
This will map the user the CI runs under corresponding to a
pre-configured mapping based on the group of the repository.

The user running the CI will not, in general, have an SSH key registered
with BBP's GitLab instance, so we rewrite git URLs to use HTTPS and the
job token GitLab passes in the $CI_JOB_TOKEN environment variable.

Co-authored-by: Olli Lupton <oliver.lupton@epfl.ch>