Improve and simplify CI.
Use a new GITLAB_PIPELINES_BRANCH variable for testing changes to this
pipeline. Pipelines that are used in the CI (i.e. coreneuron,
touchdetector, spykfunc and their descendants like blueconfigs) must
propagate this explicitly. This allows changes to be tested in
grandchild pipelines such as blueconfigs. The default value (main) must
be set as a CI/CD variable in all such projects.