MPI_DSM_DISTRIBUTE: set to 0 and write an essay.
Setting `SLURM_OVERLAP=1` and `MPI_DSM_DISTRIBUTE=0` tries to make `mpiexec -n <N>` do something sensible when used inside a CI job that runs on BB5.

Because of the way the BB5 runner is set up, the CI script (i.e. the lines you put inside the `before_script` and `script` YAML blocks) is run inside the job allocation using `srun` and occupies all the resources.
If we want to spawn another job from that CI script, we have to use `--overlap` so that it is allowed to run on those already-occupied occupied resources.
This means that jobs spawned from the CI script are also allocated all of the resources in the job allocation.

If the spawned child job is an MPI application then by default HPE MPI will pin the MPI ranks to cores starting from the beginning of the allocated resources.
Because all child jobs are allocated all of the resources, this means that the child ranks are all pinned to the same cores.
Setting `MPI_DSM_DISTRIBUTE=0` disables this pinning and lets the OS scheduler spread the spawned jobs over the available cores.
This might not work well if the allocated resources span multiple nodes.

This is definitely a hack. It seems like a better option would to execute the CI script in a way that does not **directly** occupy resources, but which still runs it on an allocated node. The natural way to do this seems to be with `sbatch`. https://ecp-ci.gitlab.io/docs/admin/jacamar/executors.html#slurm-sbatch is an example of GitLab-Slurm integration that takes this direction.

Two common use-cases are:

**Building**
```yaml
job:
  script:
   # This should execute on the allocated resources, even though there is no child srun/mpiexec
   - ninja -j 16
```
**Testing**
```yaml
job:
  script:
   # These should execute concurrently on disjoint resources if resources are allocated for >=8 tasks
   - mpiexec -n 4 sleep 10 &
   - mpiexec -n 4 sleep 10 &
   - wait
```